管理ページは全てログインが必要なように修正

+++ routes/web.php
@@ -26,14 +26,14 @@ Route::post('contact', [ContactController::class, 'sendmail']) ;
 Route::get('/contact/complete', [ContactController::class, 'complete'])->name('contact.complete');
 
 Route::get('/admin/blogs', [AdminBlogController::class, 'index'])->name('admin.blogs.index')->middleware('auth');
-Route::get('/admin/blogs/create', [AdminBlogController::class, 'create'])->name('admin.blogs.create');
-Route::post('/admin/blogs/store', [AdminBlogController::class, 'store'])->name('admin.blogs.store');
-Route::get('/admin/blogs/{blog}', [AdminBlogController::class, 'edit'])->name('admin.blogs.edit');
-Route::put('/admin/blogs/{blog}', [AdminBlogController::class, 'update'])->name('admin.blogs.update');
-Route::delete('/admin/blogs/{blog}', [AdminBlogController::class, 'destroy'])->name('admin.blogs.delete');
+Route::get('/admin/blogs/create', [AdminBlogController::class, 'create'])->name('admin.blogs.create')->middleware('auth');
+Route::post('/admin/blogs/store', [AdminBlogController::class, 'store'])->name('admin.blogs.store')->middleware('auth');
+Route::get('/admin/blogs/{blog}', [AdminBlogController::class, 'edit'])->name('admin.blogs.edit')->middleware('auth');
+Route::put('/admin/blogs/{blog}', [AdminBlogController::class, 'update'])->name('admin.blogs.update')->middleware('auth');
+Route::delete('/admin/blogs/{blog}', [AdminBlogController::class, 'destroy'])->name('admin.blogs.delete')->middleware('auth');
 
-Route::get('/admin/users/create', [AdminUserController::class, 'create'])->name('admin.users.create');
-Route::post('/admin/users/store', [AdminUserController::class, 'store'])->name('admin.users.store');
+Route::get('/admin/users/create', [AdminUserController::class, 'create'])->name('admin.users.create')->middleware('auth');
+Route::post('/admin/users/store', [AdminUserController::class, 'store'])->name('admin.users.store')->middleware('auth');
 
 Route::get('/admin/login', [AuthController::class, 'showLoginForm'])->name('admin.login');
 Route::post('/admin/login', [AuthController::class, 'login']);
